+++ android/anc/src/main/java/org/smartregister/fhircore/anc/util/SearchExtension.kt
@@ -21,21 +21,31 @@ import ca.uhn.fhir.rest.gclient.StringClientParam
 import ca.uhn.fhir.rest.gclient.TokenClientParam
 import com.google.android.fhir.search.Search
 import com.google.android.fhir.search.StringFilterModifier
+import org.hl7.fhir.r4.model.CodeableConcept
+import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.Enumerations
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.ResourceType
-import org.smartregister.fhircore.anc.sdk.QuestionnaireUtils.asCodeableConcept
+
+fun Coding.asCodeableConcept() =
+  CodeableConcept().apply {
+    addCoding(this@asCodeableConcept)
+    text = this@asCodeableConcept.display
+  }
 
 fun Search.filterByPatient(reference: ReferenceClientParam, patientId: String) {
-  filter(reference) { this.value = "${ResourceType.Patient.name}/$patientId" }
+  filter(reference, { value = "${ResourceType.Patient.name}/$patientId" })
 }
 
 fun Search.filterByPatientName(name: String?) {
   if (name?.isNotBlank() == true) {
-    filter(Patient.NAME) {
-      value = name.trim()
-      modifier = StringFilterModifier.CONTAINS
-    }
+    filter(
+      Patient.NAME,
+      {
+        modifier = StringFilterModifier.CONTAINS
+        value = name.trim()
+      }
+    )
   }
 }
 
@@ -43,10 +53,13 @@ fun Search.filterBy(filter: SearchFilter) {
   when (filter.filterType) {
     Enumerations.SearchParamType.TOKEN -> filterToken(filter)
     Enumerations.SearchParamType.STRING ->
-      filter(StringClientParam(filter.key)) {
-        this.modifier = StringFilterModifier.MATCHES_EXACTLY
-        this.value = filter.valueString!!
-      }
+      filter(
+        StringClientParam(filter.key),
+        {
+          modifier = StringFilterModifier.MATCHES_EXACTLY
+          value = filter.valueString!!
+        }
+      )
     else ->
       throw UnsupportedOperationException("Can not apply ${filter.filterType} as search filter")
   }
@@ -55,9 +68,13 @@ fun Search.filterBy(filter: SearchFilter) {
 fun Search.filterToken(filter: SearchFilter) {
   // TODO TokenFilter in SDK is not fully implemented and ignores all types but Coding
   when (filter.valueType) {
-    Enumerations.DataType.CODING -> filter(TokenClientParam(filter.key), filter.valueCoding!!)
+    Enumerations.DataType.CODING ->
+      filter(TokenClientParam(filter.key), { value = of(filter.valueCoding!!) })
     Enumerations.DataType.CODEABLECONCEPT ->
-      filter(TokenClientParam(filter.key), filter.valueCoding!!.asCodeableConcept()!!)
+      filter(
+        TokenClientParam(filter.key),
+        { value = of(filter.valueCoding!!.asCodeableConcept()!!) }
+      )
     else ->
       throw UnsupportedOperationException("SDK does not support value type ${filter.valueType}")
   }