+++ android/engine/src/main/java/org/smartregister/fhircore/engine/util/extension/ApplicationExtension.kt
@@ -23,6 +23,7 @@ import com.google.android.fhir.db.ResourceNotFoundException
 import com.google.android.fhir.search.Order
 import com.google.android.fhir.search.StringFilterModifier
 import com.google.android.fhir.search.count
+import com.google.android.fhir.search.getQuery
 import com.google.android.fhir.search.search
 import com.google.android.fhir.sync.ResourceSyncParams
 import com.google.android.fhir.sync.State
@@ -69,12 +70,18 @@ suspend fun FhirEngine.searchActivePatients(
   loadAll: Boolean = false
 ) =
   this.search<Patient> {
-    filter(Patient.ACTIVE, true)
+    apply { filter(Patient.ACTIVE, { value = of(true) }) }.getQuery()
     if (query.isNotBlank()) {
-      filter(Patient.NAME) {
-        modifier = StringFilterModifier.CONTAINS
-        value = query.trim()
-      }
+      apply {
+          filter(
+            Patient.NAME,
+            {
+              modifier = StringFilterModifier.CONTAINS
+              value = query.trim()
+            }
+          )
+        }
+        .getQuery()
     }
     sort(Patient.NAME, Order.ASCENDING)
     count =
@@ -84,7 +91,7 @@ suspend fun FhirEngine.searchActivePatients(
   }
 
 suspend fun FhirEngine.countActivePatients(): Long =
-  this.count<Patient> { filter(Patient.ACTIVE, true) }
+  this.count<Patient> { apply { filter(Patient.ACTIVE, { value = of(true) }) }.getQuery(true) }
 
 suspend inline fun <reified T : Resource> FhirEngine.loadResource(resourceId: String): T? {
   return try {
@@ -97,7 +104,7 @@ suspend inline fun <reified T : Resource> FhirEngine.loadResource(resourceId: St
 suspend fun FhirEngine.loadRelatedPersons(patientId: String): List<RelatedPerson>? {
   return try {
     this@loadRelatedPersons.search {
-      filter(RelatedPerson.PATIENT) { value = "Patient/$patientId" }
+      apply { filter(RelatedPerson.PATIENT, { value = "Patient/$patientId" }) }.getQuery()
     }
   } catch (resourceNotFoundException: ResourceNotFoundException) {
     null
@@ -107,7 +114,7 @@ suspend fun FhirEngine.loadRelatedPersons(patientId: String): List<RelatedPerson
 suspend fun FhirEngine.loadPatientImmunizations(patientId: String): List<Immunization>? {
   return try {
     this@loadPatientImmunizations.search {
-      filter(Immunization.PATIENT) { value = "Patient/$patientId" }
+      apply { filter(Immunization.PATIENT, { value = "Patient/$patientId" }) }.getQuery()
     }
   } catch (resourceNotFoundException: ResourceNotFoundException) {
     null