+++ CHANGELOG.md
@@ -55,12 +55,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Fixed multiple entries of same test results on quest patient detail screen
 - Quest | Fixed mislabeling of questionnaire responses on quest patient detail screen
 - Quest | Fix patient registration with estimated age/dob
+- Quest | Fix patient register data list items display
 - Engine | Fixed app crash when wrong appId is provided
 - Engine | Fixed login error message
 
 ### Changed
 - EIR | Updated overdue trigger flow for Vaccine Due date
 - EIR | Updated on save vaccine received alert dialogue
+- ANC | Removed custom extraction handling and moved Family/ANC forms to structure map
 
 - Quest | Updated Sync to sync user specific fhir resources using Search Parameters from assets